feat: add roms for testing
This commit adds a set of roms compiled from:
https://www.zophar.net/pdroms/chip8.html

Listed/Published as:
Files such as emulators, hacks, utilities and
other programs that can be used to modify or
use copyrighted content themselves are not
illegal, and they never will be illegal.
The emulation scene and this site has
been alive for over two decades, and we
will forever be here.

Under: https://www.zophar.net/legal.html